+++ src/Analysis/Linting/UnknownGlobalFunctionAnalyzer.php
@@ -2,12 +2,11 @@
 
 namespace PhpIntegrator\Analysis\Linting;
 
+use PhpIntegrator\Analysis\FunctionCallNodeFqsenDeterminer;
 use PhpIntegrator\Analysis\GlobalFunctionExistanceCheckerInterface;
 
 use PhpIntegrator\Analysis\Visiting\NamespaceAttachingVisitor;
 use PhpIntegrator\Analysis\Visiting\GlobalFunctionUsageFetchingVisitor;
-
-use PhpIntegrator\Utility\NodeHelpers;
 
 /**
  * Looks for unknown global function names (i.e. used during calls).
@@ -60,37 +59,18 @@ class UnknownGlobalFunctionAnalyzer implements AnalyzerInterface
 
         $unknownGlobalFunctions = [];
 
+        // TODO: Inject this.
+        $determiner = new FunctionCallNodeFqsenDeterminer($this->globalFunctionExistanceChecker);
+
         foreach ($globalFunctions as $node) {
-            $name = NodeHelpers::fetchClassName($node->name->getAttribute('resolvedName'));
+            $fqsen = $determiner->determine($node);
 
-            $namespaceNode = $node->getAttribute('namespace');
-            $namespace = null;
-
-            if ($namespaceNode !== null) {
-                $namespace = NodeHelpers::fetchClassName($namespaceNode);
-            }
-
-            if ($this->globalFunctionExistanceChecker->doesGlobalFunctionExist($name)) {
+            if ($this->globalFunctionExistanceChecker->doesGlobalFunctionExist($fqsen)) {
                 continue;
-            } elseif ($node->name->isUnqualified()) {
-                // Unqualified global function calls, such as "array_walk", could refer to "array_walk" in the current
-                // namespace (e.g. "\A\array_walk") or, if not present in the current namespace, the root namespace
-                // (e.g. "\array_walk").
-                $fqcnForCurrentNamespace = '\\' . $namespace . '\\' . $name;
-
-                if ($this->globalFunctionExistanceChecker->doesGlobalFunctionExist($fqcnForCurrentNamespace)) {
-                    continue;
-                }
-
-                $fqcnForRootNamespace = '\\' . $name;
-
-                if ($this->globalFunctionExistanceChecker->doesGlobalFunctionExist($fqcnForRootNamespace)) {
-                    continue;
-                }
             }
 
             $unknownGlobalFunctions[] = [
-                'name'  => $name,
+                'name'  => $fqsen,
                 'start' => $node->getAttribute('startFilePos') ? $node->getAttribute('startFilePos')   : null,
                 'end'   => $node->getAttribute('endFilePos')   ? $node->getAttribute('endFilePos') + 1 : null
             ];